separated useHalls hook and Hall context
